importers/requirements.txt
httpx==0.27.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
azure-storage-blob==12.19.1
chardet==5.2.0
EbookLib==0.18
python-slugify==8.0.4
aiofiles==23.2.1
